Move textual representation functions to platform module
